double -> float

+++ CudaLearning-2/vec3.h
@@ -94,23 +94,23 @@ __host__ __device__ inline vec3 operator/(const vec3& v1, const vec3& v2) {
 
 // ³ËÒÔ³£”µ
 
-__host__ __device__ inline vec3 operator*(double t, const vec3& v)
+__host__ __device__ inline vec3 operator*(float t, const vec3& v)
 {
 	return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-__host__ __device__ inline vec3 operator*(const vec3& v, double t)
+__host__ __device__ inline vec3 operator*(const vec3& v, float t)
 {
 	return t * v;
 }
 
-__host__ __device__ inline vec3 operator/(vec3 v, double t)
+__host__ __device__ inline vec3 operator/(vec3 v, float t)
 {
 	return (1 / t) * v;
 }
 
 // üc³Ë
-__host__ __device__ inline double dot(const vec3& u, const vec3& v)
+__host__ __device__ inline float dot(const vec3& u, const vec3& v)
 {
 	return u.e[0] * v.e[0] + u.e[1] * v.e[1] + u.e[2] * v.e[2];
 }
